apimanager/requirements.txt
asgiref==3.7.2
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==4.1.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.7
decorator==5.1.1
Django==3.2.25
djangorestframework==3.15.1
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-bigquery==3.22.0
google-cloud-core==2.4.1
google-cloud-pubsub==2.21.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.2
grpcio-status==1.62.2
idna==3.7
importlib-metadata==6.7.0
ipdb==0.13.13
ipython==7.34.0
jedi==0.19.1
matplotlib-inline==0.1.6
packaging==24.0
paramiko==3.4.0
parso==0.8.4
pickleshare==0.7.5
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.24.4
psycopg2-binary==2.9.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pygments==2.17.2
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==227
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
traitlets==5.9.0
typing-extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.13
websocket-client==1.6.1
zipp==3.15.0